+++ test/libsolidity/semanticTests/array/function_array_cross_calls.sol
@@ -45,6 +45,6 @@ contract C {
 // compileViaYul: also
 // ----
 // test() -> 5, 6, 7
-// gas irOptimized: 345955
+// gas irOptimized: 345929
 // gas legacy: 508437
 // gas legacyOptimized: 309013