+++ test/libsolidity/semanticTests/array/push/push_no_args_bytes.sol
@@ -23,7 +23,7 @@ contract C {
 // ----
 // l() -> 0
 // g(uint256): 70 ->
-// gas irOptimized: 428829
+// gas irOptimized: 429179
 // gas legacy: 419791
 // gas legacyOptimized: 415408
 // l() -> 70